Stop verifying project on activation.
