+++ packages/core/src/amazonqGumby/activation.ts
@@ -14,7 +14,6 @@ import { CodeTransformTelemetryState } from './telemetry/codeTransformTelemetryS
 import { telemetry } from '../shared/telemetry/telemetry'
 import { CancelActionPositions } from './telemetry/codeTransformTelemetry'
 import { AuthUtil } from '../codewhisperer/util/authUtil'
-import { validateAndLogProjectDetails } from '../codewhisperer/service/transformByQ/transformProjectValidationHandler'
 
 export async function activate(context: ExtContext) {
     void vscode.commands.executeCommand('setContext', 'gumby.wasQCodeTransformationUsed', false)
@@ -71,7 +70,5 @@ export async function activate(context: ExtContext) {
 
             workspaceChangeEvent
         )
-
-        await validateAndLogProjectDetails()
     }
 }
