+++ packages/core/src/codewhisperer/service/transformByQ/transformProjectValidationHandler.ts
@@ -7,12 +7,9 @@ import { getLogger } from '../../../shared/logger'
 import * as CodeWhispererConstants from '../../models/constants'
 import * as vscode from 'vscode'
 import { spawnSync } from 'child_process' // Consider using ChildProcess once we finalize all spawnSync calls
-import { CodeTransformJavaSourceVersionsAllowed, telemetry } from '../../../shared/telemetry/telemetry'
+import { telemetry } from '../../../shared/telemetry/telemetry'
 import { CodeTransformTelemetryState } from '../../../amazonqGumby/telemetry/codeTransformTelemetryState'
-import {
-    javapOutputToTelemetryValue,
-    JDKToTelemetryValue,
-} from '../../../amazonqGumby/telemetry/codeTransformTelemetry'
+import { javapOutputToTelemetryValue } from '../../../amazonqGumby/telemetry/codeTransformTelemetry'
 import { MetadataResult } from '../../../shared/telemetry/telemetryClient'
 import {
     NoJavaProjectsFoundError,
@@ -20,31 +17,6 @@ import {
     NoOpenProjectsError,
 } from '../../../amazonqGumby/errors'
 import { checkBuildSystem } from './transformFileHandler'
-
-// log project details silently
-export async function validateAndLogProjectDetails() {
-    let telemetryJavaVersion = JDKToTelemetryValue(JDKVersion.UNSUPPORTED) as CodeTransformJavaSourceVersionsAllowed
-    let errorCode = undefined
-    try {
-        const openProjects = await getOpenProjects()
-        const validProjects = await validateOpenProjects(openProjects, true)
-        if (validProjects.length > 0) {
-            // validProjects[0].JDKVersion will be undefined if javap errors out or no .class files found, so call it UNSUPPORTED
-            const javaVersion = validProjects[0].JDKVersion ?? JDKVersion.UNSUPPORTED
-            telemetryJavaVersion = JDKToTelemetryValue(javaVersion) as CodeTransformJavaSourceVersionsAllowed
-        }
-    } catch (err: any) {
-        errorCode = err.code
-    } finally {
-        telemetry.codeTransform_projectDetails.emit({
-            passive: true,
-            codeTransformSessionId: CodeTransformTelemetryState.instance.getSessionId(),
-            codeTransformLocalJavaVersion: telemetryJavaVersion,
-            result: errorCode ? MetadataResult.Fail : MetadataResult.Pass,
-            reason: errorCode,
-        })
-    }
-}
 
 export async function getOpenProjects(): Promise<TransformationCandidateProject[]> {
     const folders = vscode.workspace.workspaceFolders
